requirements.txt
pandas
#matplotlib
tweepy
python-dotenv
psycopg2-binary
#tqdm